+++ agent-control/tests/on_host/scenarios.rs
@@ -2,6 +2,7 @@ mod attributes;
 mod empty_config;
 mod filesystem_ops;
 mod health_check;
+mod id_vars;
 mod invalid_remote_config;
 mod multiple_executables;
 mod opamp;
